move plotting into plotting.py
